+++ src/graphq/resolvers.ts
@@ -1,0 +1,6 @@
+
+export default {
+  Query: {
+    hello: (): string => 'Hello world!'
+  },
+}
