--- src/schema/resolvers/index.ts
@@ -1,9 +1,0 @@
-
-
-import hello from './query/hello'
-
-export default {
-  Query: {
-    hello,
-  },
-}
