--- src/schema/resolvers/query/hello.ts
@@ -1,1 +1,0 @@
-export default (): string => 'Hello world!'
